requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu==1.7.1
langchain_google_genai
langchain_community
